server/requirements.txt
# --- CORE API & SERVER ---
fastapi==0.135.1
uvicorn==0.41.0
websockets==12.0
python-dotenv
starlette
pydantic

# --- DATABASE & SUPABASE ---
SQLAlchemy==2.0.30
alembic
asyncpg
pgvector
supabase==2.4.0

# --- AI & COMPUTER VISION (GPU & INSIGHTFACE SOTA) ---
numpy==1.26.4
onnxruntime-gpu==1.23.2
opencv-python==4.9.0.80
insightface==0.7.3

# --- MACHINE LEARNING & UTILS ---
torch==2.2.1
torchvision==0.17.1
albumentations
pandas
openpyxl
scikit-learn
scikit-image
tqdm
colorama
PyJWT

# STEP1: pip install numpy==1.26.4 opencv-python==4.9.0.80 onnxruntime-gpu==1.23.2 scipy==1.11.4 scikit-learn==1.3.2 scikit-image==0.22.0
# STEP 2: pip install torch==2.2.1 torchvision==0.17.1 insightface==0.7.3 albumentations pandas openpyxl tqdm colorama
# STEP 3: pip install fastapi==0.135.1 uvicorn==0.41.0 websockets==12.0 SQLAlchemy==2.0.30 supabase==2.4.0 asyncpg pgvector alembic pydantic python-dotenv